refactor(types): типы от фреймворков OneScript вынесены из инференсера
Четвёртая группа знаний, не относящихся к выводу типов: тип, который переменной задаёт не
код и не комментарий, а фреймворк. Их два — «ОСень» (внедрение по `&Пластилин`) и
библиотека extends (наследование по `&Расширяет`, тип поля-держателя родителя, включая
неявное `_ОбъектРодитель`).

Вынесено в `OScriptFrameworkTypes`. Вместе с группой уехали обе внешние зависимости —
вывод компонентов «ОСени» и разбор аннотаций наследования: за её пределами они не
использовались.

Разрешение неявного поля родителя перестало возвращать `Optional`: пустой набор типов
означает ровно то же, а звать его снаружи стало проще.

Инференсер: 1350 строк → 1287, от 1650 на develop.
